Set node TTL to 1 if explicitly stopping container
Nodes will be removed from the cluster if their TTL in nodes/<ip/hostname> runs
out (default to 60 seconds).  If we're stopping the watch daemon (which also
extends the TTL), we set the TTL to 1.  In testing, use a fixed docker hostname
instead of an IP, so that rejoining works properly.